% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{Cluster Gene Sets}
\format{A list of 10 clusters formed through hierarchical clustering of the 
  4000 probes described in \code{\link{tb.expr}}. The probes were clustered
  after centering and scaling each of the rows.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
clusters
}
\description{
A list of clusters
}
\keyword{datasets}
