% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{des.info}
\alias{des.info}
\title{Design information object}
\format{A list of length 16. See \code{\link{desInfo}} for details.}
\source{
\code{\link{desInfo}}
}
\usage{
des.info
}
\description{
A list generated by \code{\link{desInfo}}. Contains matched design and
expression datasets, along with defined design parameters.
}
\keyword{datasets}
