% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDendrograms.R
\name{genDendrograms}
\alias{genDendrograms}
\title{Normalization and hierarchical clustering}
\usage{
genDendrograms(design_info)
}
\arguments{
\item{design_info}{list generated from \code{desInfo}}
}
\value{
Returns dendrogram objects obtained through hierarchical clustering
  of the various normalized expression datasets.
}
\description{
Performs hierarchical clustering and various normalizations of expression
matrix
}
\details{
This function performs various normalizations of the expression
  data, depending on the study design and the parameters defined in
  \code{\link{desInfo}}. For all study designs, the data is normalized to the
  mean of all the samples. For non-logitudinal studies with controls, an
  additional normalization to the mean of the controls is performed. For
  longitudinal studies, baseline normalization is performed (each subject's
  baseline is subtracted out from later time points), along with separate
  normalization of baseline samples only (proceeding as with non-longitudinal
  studies). For each normalized expression matrix, hierarchical clustering of
  the rows is performed.
}
\examples{
# des.info object is obtained using the desInfo function
data(des.info)
dendros <- genDendrograms(des.info)
}
