% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.symbols}
\alias{gene.symbols}
\title{Gene Symbols}
\format{A vector of 1000 gene symbols}
\source{
Illumina
}
\usage{
gene.symbols
}
\description{
A vector of gene symbols to match the probes ids in expression
}
\keyword{datasets}
