% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBartFile.R
\name{genFile}
\alias{genFile}
\alias{updateFile}
\title{Generate and Update BART files}
\usage{
genFile(design_info, model_results = NULL, module_maps = NULL,
  dendros = NULL, qusage_results = NULL, roast_results = NULL,
  corr_results = NULL, illumina = TRUE)

updateFile(load.path = NULL, output.path = NULL, design_info = NULL,
  model_results = NULL, module_maps = NULL, dendros = NULL,
  qusage_results = NULL, roast_results = NULL, corr_results = NULL,
  illumina = NULL)
}
\arguments{
\item{design_info}{list generated from \code{desInfo}}

\item{model_results}{Default is NULL. Object generated from function
\code{genModelResults}}

\item{module_maps}{Default is NULL. Object generated from function
\code{genModules}}

\item{dendros}{Default is NULL. Object generated from function
\code{genDendrograms}}

\item{qusage_results}{Default is NULL. Object generated from function
\code{qBart}}

\item{roast_results}{Default is NULL. Object generated from function
\code{rBart}}

\item{corr_results}{Default is NULL. Object generated from function
\code{crossCorr}}

\item{illumina}{Logical. Are the probe IDs Illumina? Default is TRUE.}

\item{load.path}{Folder path of BART file that needs to be updated.}

\item{output.path}{Folder path to save updated BART file. If NULL (default),
updated BART file will override the old one.}
}
\description{
Generate and/or update BART file that can be directly uploaded into the BART
app
}
\details{
\code{genFile} creates a formatted BART file (Unsupervised.RData)
  based on input components that the user provides. The function then saves
  the file in a folder named after the project name that was defined in the
  function \code{\link{desInfo}}. The folder is created in the current
  workign directory. All function inputs are the objects generated from
  running the other available functions in \code{genBART}.

\code{updateFile} takes an existing BART file and allows the user to
  easily add and/or revise components. The user can override the old BART
  file by leaving \code{output.path} as NULL or give a new path in which to
  save the updated file.
}
\examples{
# des.info and mod.results objects are obtained using the desInfo and genModelResults functions respectively
data(des.info)
data(mod.results)

# gene sets and data for correlations
data(modules)
data(flow)
data(mod.as)

# generate module maps and cluster probes
mods <- genModules(des.info, modules)
dendros <- genDendrograms(des.info)

# create BART file
genFile(design_info = des.info, module_maps = mods, dendros = dendros)

# run qusage and correlations
qus <- qBart(mod.results, modules)
corrs <- crossCorr(x = mod.as, y = flow.dat, by = time, by_name = "days", description = "Mod.Act.Score vs Flow",
                   x_var = "Mod.Act.Score", y_var = "Flow",method = "spearman")

# update and override BART file by adding qusage results and correlations
updateFile(load.path = paste(getwd(), "/", project_name, " Pipeline/", sep = ""), qusage_results = qus, corr_results = corrs)
}
