% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model.results}
\alias{model.results}
\title{Model results object}
\format{A list of length 3. See \code{\link{genModelResults}} for details.}
\source{
\code{\link{genModelResults}}
}
\usage{
model.results
}
\description{
A list generated by \code{\link{genModelResults}}.
}
\keyword{datasets}
