% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module.as}
\alias{module.as}
\title{Module Activity Scores}
\format{A data frame with 194 rows of samples and 260 columns of baylor
  modules.}
\source{
baylor modules and 
  \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
module.as
}
\description{
This is a dataset containing module activity scores for cynomolgus macaques 
infected with M. tuberculosis. For a given module, monkey, and time point, 
each score was calculated by summing the difference in probe-level 
post-infection expression values and the average basline expression value for
all probes within the module. This sum was then divided by the number of 
probes within each module plus the number of samples within each time point.
}
\keyword{datasets}
