% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modules}
\alias{modules}
\title{Baylor Modules}
\format{A large list containing 260 gene sets.}
\source{
baylor modules
}
\usage{
modules
}
\description{
A list containing the baylor modules (gene sets)
}
\keyword{datasets}
