% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetTesting.R
\name{qBart}
\alias{qBart}
\title{Run Qusage algorithm using gene level statistics}
\usage{
qBart(object, gene_sets)
}
\arguments{
\item{object}{object generated from \code{genModelResults}}

\item{gene_sets}{list of gene_sets}
}
\value{
\code{qusage_results} tall formatted matrix of results

\code{lowerCI} Matrix of gene level lower 95\% confidence intervals

\code{upperCI} Matrix gene level upper 95\% confidence intervals
}
\description{
Run Qusage algorithm using gene level statistics
}
\details{
This function takes the gene level comparison estimates and test
  statistics contained in the object produced from
  \code{\link{genModelResults}} and runs the Qusage algorithm across all of
  the comparisons. The VIFs are estimated using the raw residuals, which are
  also contained in the output of \code{\link{genModelResults}}.
}
\examples{
# mod.results object is obtained using the genModelResults function
data(mod.results)
data(modules)
qus <- qBart(mod.results, modules)
}
