% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBART.R
\name{runBart}
\alias{runBart}
\title{Run BART Shiny App}
\usage{
runBart()
}
\description{
This function runs the BART shiny app
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runBart()
}
}
