% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.design}
\alias{tb.design}
\title{Design file for cynomolgus macaques infected with M. tuberculosis}
\format{A data frame with 436 rows and 7 variables.
\describe{
  \item{columnname}{names matching column names in expression dataset}
  \item{monkey_id}{id assigned to annotate each monkey}
  \item{timepoint}{times at which samples were drawn, in days}
  \item{timerange}{groups time points into stages}
  \item{sample_id}{id assigned to annotate each sample}
  \item{clinical_status}{denotes if disease is active or latent}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
tb.design
}
\description{
A dataset containing sample information such as time point and clinical status.
}
\keyword{datasets}
