% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.flow}
\alias{tb.flow}
\title{Flow data on cynomolgus macaques infected with M. tuberculosis}
\format{A data frame with 13 rows of flow variables and 213 columns of
  samples}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
tb.flow
}
\description{
A dataset containing flow variables of absolute counts and percentages.
}
\keyword{datasets}
