% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDendrograms.R
\name{genDendrograms}
\alias{genDendrograms}
\title{Normalization and hierarchical clustering}
\usage{
genDendrograms(design_info)
}
\arguments{
\item{design_info}{list generated from \code{desInfo}}
}
\value{
Returns dendrogram objects obtained through hierarchical clustering 
  of the various normalized expression datasets.
}
\description{
Performs hierarchical clustering and various normalizations of expression 
matrix
}
\details{
This function performs various normalizations of the expression 
  data, depending on the study design and the parameters defined in 
  \code{\link{desInfo}}. For all study designs, the data is normalized to the
  mean of all the samples. For non-logitudinal studies with controls, an 
  additional normalization to the mean of the controls is performed. For 
  longitudinal studies, baseline normalization is performed (each subject's 
  baseline is subtracted out from later time points), along with separate 
  normalization of baseline samples only (proceeding as with non-longitudinal
  studies). For each normalized expression matrix, hierarchical clustering of
  the rows is performed.
}
\examples{
# Example data
data(tb.expr)
data(tb.design)

# Use first 100 probes to demonstrate
dat <- tb.expr[1:100,]

# Create desInfo object
des.info <- desInfo(y = dat, design = tb.design, data_type = "micro", 
                    columnname = "columnname", long = TRUE, sample_id = "sample_id",
                    patient_id = "monkey_id", time_var = "timepoint",
                    baseline_var = "timepoint", baseline_val = 0, 
                    responder_var = "clinical_status", project_name = "TB")

# Normalize and cluster data
dendros <- genDendrograms(des.info)
}
