% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genModelResults.R
\name{genModelResults}
\alias{genModelResults}
\title{Generate formatted results file}
\usage{
genModelResults(design_info, object, lm_Fit, method = "limma",
  comp_names = NULL, var_names = NULL, var_symbols = NULL)
}
\arguments{
\item{design_info}{list generated from \code{\link{desInfo}}}

\item{object}{result object generated from \code{limma}, \code{DESeq2}, or 
\code{edgeR}}

\item{lm_Fit}{linear model fit object generated from \code{limma}, 
\code{DESeq2}, or \code{edgeR}}

\item{method}{Which method is being used? ("limma","deseq2","edgeR")}

\item{comp_names}{comp_names=NULL is the default. A vector of comparison 
(contrast) names}

\item{var_names}{var_names=NULL is the default. Optional vector of variable 
names. Otherwise, the rownames of the expression dataframe are used}

\item{var_symbols}{var_symbols=NULL is the default. Optional vector of 
secondary annotations for the variables. Otherwise, the rownames of the 
expression dataframe are used}
}
\value{
\code{data_type} String denoting the type of data that was analyzed

\code{results} The formatted results file

\code{resids} Matrix of residuals. Returned only if data_type="micro"
  or "rna". Used to estimate VIFs when running the QUSAGE algorithim in 
  \code{\link{qBart}}
}
\description{
Generate formatted results file from completed differential analysis in 
limma, DESeq2, or edgeR
}
\details{
This function formats the results obtained from running differential
  analysis in either one of \code{limma}, \code{DESeq2}, or \code{edgeR}. The
  parameter \code{object} accepts as input a results object from functions 
  \code{\link[limma]{eBayes}} (limma), \code{\link[DESeq2]{results}} 
  (DESeq2), \code{\link[edgeR]{glmLRT}} or \code{\link[edgeR]{glmQLFTest}} 
  (edgeR). The parameter \code{lm_Fit} accepts a fitted model object from 
  functions \code{\link[limma]{lmFit}} (limma), \code{\link[DESeq2]{DESeq}} 
  (DESeq2), \code{\link[edgeR]{glmFit}} or \code{\link[edgeR]{glmQLFit}} 
  (edgeR).
}
\examples{
# Using example dataset in genBART package
data(tb.expr)
data(tb.design)

# Gene symbols
data(gene.symbols)

# des.info object is obtained using the desInfo function
data(des.info)

# Generate lmFit and eBayes (limma) objects needed for genModelResults
tb.design$Group <- paste(tb.design$clinical_status,tb.design$timepoint,
                         sep = "")
grp <- factor(tb.design$Group)
design2 <- model.matrix(~0+grp)
colnames(design2) <- levels(grp)
dupcor <- limma::duplicateCorrelation(tb.expr, design2, block = tb.design$monkey_id)
fit <- limma::lmFit(tb.expr, design2, block = tb.design$monkey_id, 
             correlation = dupcor$consensus.correlation)
contrasts <- limma::makeContrasts(A_20vsPre = Active20-Active0, A_30vsPre = Active30-Active0,
                           A_42vsPre = Active42-Active0, A_56vsPre = Active56-Active0,
                           levels=design2)
fit2 <- limma::contrasts.fit(fit, contrasts)
fit2 <- limma::eBayes(fit2, trend = FALSE)

mod.results <- genModelResults(design_info = des.info, object = fit2, lm_Fit = fit, 
                               method = "limma", var_symbols = gene.symbols)
}
