% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetTests.R
\name{rBart}
\alias{rBart}
\title{Run ROAST method and format results for BART}
\usage{
rBart(meta, gene.sets, design, contrast, block = NULL, correlation = NULL,
  gene.weights = NULL, var.prior = NULL, df.prior = NULL, nrot = 999)
}
\arguments{
\item{meta}{list generated from \code{metaData}}

\item{gene.sets}{list of gene sets}

\item{design}{design matrix}

\item{contrast}{numeric matrix with rows corresponding to coefficients and
columns containing contrasts. May be a vector if there is only one
contrast.}

\item{block}{vector or factor specifying a blocking variable on the samples}

\item{correlation}{the inter-duplicate or inter-technical replicate
correlation}

\item{gene.weights}{numeric vector of directional (positive or negative)
probewise weights. Must have length equal to the number of rows in the
expression dataset.}

\item{var.prior}{prior value for residual variances. If not provided, this is
estimated from all the data using squeezeVar.}

\item{df.prior}{prior degrees of freedom for residual variances. If not
provided, this is estimated using squeezeVar.}

\item{nrot}{number of rotations used to compute the p-values.}
}
\description{
Run ROAST method and format results for BART
}
\details{
This function runs the ROAST gene set tests across all summary set
  statistics ("mean", "floormean", "mean50", and "msq") and returns formatted
  results for BART.
}
