% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.expr}
\alias{tb.expr}
\title{Microarray dataset on cynomolgus macaques infected with M. tuberculosis}
\format{A data frame with 1000 rows and 436 columns, where each row is a 
probe and each column a sample.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
tb.expr
}
\description{
A dataset containing gene expression data on 38 monkeys at various time 
points.
}
\keyword{datasets}
