% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tb.flow}
\alias{tb.flow}
\title{Flow dataset on cynomolgus macaques infected with M. tuberculosis}
\format{A data frame with 194 rows of samples and 13 flow variables.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84152}
}
\usage{
tb.flow
}
\description{
A dataset containing flow absolute count and percentage data
}
\keyword{datasets}
